Filter MCP servers (#65)
Anyone can implement an MCP Server and submit it to a registry, much
like building a container image and push it to Docker Hub.
However, not all these servers are stable, bug-free, up-to-date,
trusted, and safe to run locally.

Add a filter to use only "official" MCP servers if users choose to.

Closes #65

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>