backend: Add agent-factory backend
Add the agent-factory backend, which provides HTTP endpoints to:
* Create agents using the agent-factory agent.
* Store metadata about agents in a database.
* Retrieve generated agents as `.zip` files

The async, long-running process of creating an agent is offloaded to a
Celery task, thus, the app needs a broker to run. The chosen broker in
this case is Redis.

Refs #90

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

Run the entire application stack with Docker Compose

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>

backend: Store agent on MinIO

Strore the generated agent files on an S3-compatible object storage. For
local deployments, use MinIO.

Refs #90

Signed-off-by: Dimitris Poulopoulos <dimitris@mozilla.ai>